# Conflicts:
#	src/backend/ci/core/dispatch-docker/biz-dispatch-docker/src/main/kotlin/com/tencent/devops/dispatch/docker/listener/AgentLessListener.kt

+++ src/backend/ci/core/stream/biz-stream/src/main/kotlin/com/tencent/devops/stream/service/StreamRequestService.kt
@@ -97,7 +97,7 @@ class StreamRequestService @Autowired constructor(
         )
         val requestList = gitRequestEventDao.getRequestsById(
             dslContext = dslContext,
-            requestIds = messageData?.map { it.messageId.toInt() }?.toSet() ?: emptySet(),
+            requestIds = messageData?.map { it.messageId.toLong() }?.toSet() ?: emptySet(),
             hasEvent = false
         )
         if (requestList.isEmpty() || count == 0) {
